440: Split user body checks into separate files
