+++ apps/backend/src/controllers/.bodyChecks/userChecks/loginRequest.check.ts
@@ -1,16 +1,14 @@
 import { getPepper, verifyPassword } from '@repo/hash';
 import {
   insertableUserObject,
-  refreshRequestUserObject,
   zodErrorToResponse400,
   type Response429,
   type User,
 } from '@repo/votura-validators';
-import { hashRefreshToken, verifyUserToken } from '../../auth/utils.js';
-import { HttpStatusCode } from '../../httpStatusCode.js';
-import { getRetryIn, recordFailedLoginAttempt } from '../../services/loginAttempt.service.js';
-import { findDBUserBy } from '../../services/users.service.js';
-import type { BodyCheckValidationError } from './bodyCheckValidationError.js';
+import { HttpStatusCode } from '../../../httpStatusCode.js';
+import { getRetryIn, recordFailedLoginAttempt } from '../../../services/loginAttempt.service.js';
+import { findDBUserBy } from '../../../services/users.service.js';
+import type { BodyCheckValidationError } from '../bodyCheckValidationError.js';
 
 export enum LoginRequestValidationErrorMessage {
   invalidCredentials = 'Invalid credentials.',
@@ -43,11 +41,6 @@ const checkLoginBlockedError = async (
   };
 };
 
-const invalidCredentialsError = {
-  status: HttpStatusCode.unauthorized,
-  message: LoginRequestValidationErrorMessage.invalidCredentials,
-};
-
 /**
  * Save a failed login attempt to the database and return either a simple invalidCredentialsError, or a too many requests error, depending on whether the user was being blocked
  * @param ipAddress
@@ -64,9 +57,13 @@ const handleFailedLogin = async (
     return loginBlockedError;
   }
 
-  return invalidCredentialsError;
+  return {
+    status: HttpStatusCode.unauthorized,
+    message: LoginRequestValidationErrorMessage.invalidCredentials,
+  };
 };
 
+// TODO: TSDoc
 export const validateLoginRequest = async (
   reqBody: unknown,
   ipAddress: string,
@@ -111,67 +108,6 @@ export const validateLoginRequest = async (
   //if (!user.verified) {
   //  return loginError.UserNotVerified; // User not verified
   //}
-
-  return user.id; // Return user ID if validation is successful
-};
-
-// ----------- Token Refresh Request Validation -----------
-export enum TokenRefreshRequestValidationErrorMessage {
-  invalidToken = 'Invalid refresh token.',
-  userNotFound = 'User not found.',
-  tokenExpired = 'Refresh token has expired.',
-}
-
-export interface TokenRefreshRequestValidationError extends BodyCheckValidationError {
-  message: TokenRefreshRequestValidationErrorMessage | string;
-}
-
-// Authenticate refresh token -- validateRefreshRequest -> error | userId
-export const validateTokenRefreshRequest = async (
-  reqBody: unknown,
-): Promise<User['id'] | TokenRefreshRequestValidationError> => {
-  const { data, error, success } = await refreshRequestUserObject.safeParseAsync(reqBody);
-  if (!success) {
-    return {
-      status: HttpStatusCode.badRequest,
-      message: zodErrorToResponse400(error).message,
-    };
-  }
-
-  // Verify refresh token
-  const decodedToken = verifyUserToken(data.refreshToken);
-  if (decodedToken === null || decodedToken.type !== 'refresh') {
-    return {
-      status: HttpStatusCode.unauthorized,
-      message: TokenRefreshRequestValidationErrorMessage.invalidToken,
-    };
-  }
-
-  // Get user and verify stored refresh token
-  const user = await findDBUserBy({ id: decodedToken.sub });
-  if (user === null) {
-    return {
-      status: HttpStatusCode.unauthorized,
-      message: TokenRefreshRequestValidationErrorMessage.userNotFound,
-    };
-  }
-
-  // Check if refresh token matches stored hash
-  const refreshTokenHash = hashRefreshToken(data.refreshToken);
-  if (user.refreshTokenHash !== refreshTokenHash) {
-    return {
-      status: HttpStatusCode.unauthorized,
-      message: TokenRefreshRequestValidationErrorMessage.invalidToken,
-    };
-  }
-
-  // Check if refresh token is expired (expiration date in payload is already checked by verifyUserToken)
-  if (user.refreshTokenExpiresAt === null || user.refreshTokenExpiresAt < new Date()) {
-    return {
-      status: HttpStatusCode.unauthorized,
-      message: TokenRefreshRequestValidationErrorMessage.tokenExpired,
-    };
-  }
 
   return user.id; // Return user ID if validation is successful
 };
