+++ apps/backend/src/controllers/voterGroups/createVoterGroup.uc.ts
@@ -11,8 +11,8 @@ import {
 import type { Request, Response } from 'express';
 import { HttpStatusCode } from '../../httpStatusCode.js';
 import { createVoterGroup as createPersistentVoterGroup } from '../../services/voterGroups.service.js';
-import { isBodyCheckValidationError } from '../bodyChecks/bodyCheckValidationError.js';
-import { validateInsertableVoterGroup } from '../bodyChecks/voterGroupChecks.js';
+import { isBodyCheckValidationError } from '../.bodyChecks/bodyCheckValidationError.js';
+import { validateInsertableVoterGroup } from '../.bodyChecks/voterGroupChecks.js';
 
 export const createVoterGroup = async (
   req: Request,
