+++ apps/backend/src/controllers/voterGroups/updateVoterGroup.uc.ts
@@ -11,8 +11,8 @@ import {
 import type { Request, Response } from 'express';
 import { HttpStatusCode } from '../../httpStatusCode.js';
 import { updateVoterGroup as updatePersistentVoterGroup } from '../../services/voterGroups.service.js';
-import { isBodyCheckValidationError } from '../bodyChecks/bodyCheckValidationError.js';
-import { validateUpdateableVoterGroup } from '../bodyChecks/voterGroupChecks.js';
+import { isBodyCheckValidationError } from '../.bodyChecks/bodyCheckValidationError.js';
+import { validateUpdateableVoterGroup } from '../.bodyChecks/voterGroupChecks.js';
 
 export const updateVoterGroup = async (
   req: Request<{ voterGroupId: SelectableVoterGroup['id'] }>,
